FEATURE: runtime_metrics to periodically send Go runtime metrics
This currently does memory stats, gc stats, goroutines, cgo calls, and file descriptors. Send using Gauge functionality. NOTE: gauges might not be the best for some of the things. For instance, the total time spent in GC would be super nice to track as a counter which would increment by the increased amount every time period. But that's for another day...